Introduced gestures.
It is now possible to use a fling, a long press and so on to trigger
actions on the whole list, like checking everything or discarding
checked items. A contextual actions menu can also be open on gesture.

As a consequence, the option menu is simplified. All those actions are
replaced by a unique entry which opens the contextual actions menu.